[MIN-354] Move dexes tests from runtime in module
